Do not delete anything if it does not exist

+++ EssentialFeedTests/Feed Cache/CodableFeedStoreTests.swift
@@ -57,6 +57,9 @@ class CodableFeedStore {
     }
 
     func deleteCachedFeed(completion: @escaping FeedStore.DeletionCompletion) {
+        guard FileManager.default.fileExists(atPath: storeURL.path) else {
+            return completion(nil)
+        }
         do {
             try FileManager.default.removeItem(at: storeURL)
             completion(nil)
